==> notebooks/04_create_pest_list.ipynb <==
# 04 - Create list of pests for the crops in the selected area

### Description
The goal of this notebook is to:
- create a dataset with the list of possible pests that affect the crops of the selected area
- provide an interface for the selection of the pests (not implemented yet)


### Inputs

- List of crops (EPPO codes) in the area of interest, created by step 02

### External resources

- External API service: EPPO API service, to provide data on which pests affect which crop

### Outputs

- dataset with pests for each crop, including pest taxonomy: eppo_crop_pest_data.csv 
- list with common names of the pests: pest_common.csv

# import modules
import pandas as pd
import requests
%matplotlib inline

## 1. Read list of crops

In step 02 of the workflow, we created a list of the crops occurring in the area of interest, from the crop classification done by CROPLandCOS service. We added the EPPO code for each crop to that list.

# read the list of crops
in_file = '../process_data/crop_data.csv'
df = pd.read_csv(in_file)

# remove nan
df = df[df['EPPO_code'].notna()]
crops = df[['EPPO_code']]

# resolve multivalues
crops = crops['EPPO_code'].str.split(",")
crops = crops.explode('EPPO_code')
crops = crops.unique()
crops
#crops = ['ZEAMX', 'MEDSA']

## 2. Get data from EPPO API service

We will use the [EPPO Global Database](https://gd.eppo.int/) to identify which pests are known for each crop. To access to the EPPO database API, it is necessary to obtain a token. You can obtain one after registering at https://data.eppo.int/.

# define EPP API service parameters
base_url = 'https://data.eppo.int/api/rest/1.0/taxon/'
service = '/pests'
token = 'b2cdcf68881f7ac1bbfb4d0acf00b945'

Insert your EDDP database token here:

token = input('Please enter the EPPO API token:')
print(token)

The next step will gather all pests that exist in the EPPO database that are identified to have a crop as 'major host' or 'host'.

%%time
# Identifies pests for whihc the crop is identified as 'major host' or 'host'.

df = pd.DataFrame()

concat_dfs = []

for crop in crops:
    # print(crop)
    url = base_url+crop+service+'?authtoken='+token
    # print(url)
    response = requests.get(url)
    data = response.json()
    # print(data)
    if (data != None):
        if 'Major host' in data:
            df = pd.DataFrame.from_dict(data['Major host'])
        if 'Host' in data:
            df = pd.concat([df, pd.DataFrame.from_dict(data['Host'])], axis = 0, ignore_index = True, sort = True)
        df['crop_eppocode'] = crop
        concat_dfs.append(df)

df_hosts = pd.concat(concat_dfs)
df_hosts.head()

# creates and saves a dataframe with crop pests by removing duplicates
pest_codes = df_hosts[['eppocode', 'crop_eppocode']].drop_duplicates()
pest_codes.to_csv('../process_data/pest_codes.csv')

# create a unique list pest eppocodes to get its taxonomy
codes = pest_codes['eppocode'].unique()
len(codes)

A total of 920 pests where identified for the crops in the area of interest.

%%time
# Get taxonomy for each pest. This can take 4 minutes to run for a list of 2000 pest codes
group_n = 1
service = '/taxonomy'
tax = pd.DataFrame()


concatenated_dfs = []

for code in codes:
    url = base_url+code+service+'?authtoken='+token
    # print(url)
    response = requests.get(url)
    data = response.json()
    if data != None:
        tax = pd.DataFrame.from_dict(data)
        tax['kingdom'] = tax['prefname'][0]
        tax['group'] = group_n
        tax['pest_eppocode'] = code
        concatenated_dfs.append(tax)
        group_n += 1

df_tax = pd.concat(concatenated_dfs)

# retain only pests caused by animals
df_tax = df_tax[df_tax['kingdom'] == 'Animalia']

# if you have previously saved the output of the previous step, you can uncomment bellow to read a file
# and skip previous steps:

# pest_codes = pd.read_csv('../process_data/clip/crop_data.csv')
# df_tax = pd.read_csv('../process_data/pest_taxonomy.csv')

# save pest taxonomy
df_tax.to_csv('../process_data/pest_taxonomy.csv')


## 3. Create table with main taxonomic levels for each pest

This step will prepare a table with higher taxonomic levels for each pest. This is necessary to determine if birds potentially use these insects as food. Many times, insects as preys of birds are not reported at species level, but at a higher taxonomic rank.

# create a table with kingdom, order (level 5 for Insecta) and 

# group the pests
pest_tax_group = df_tax.groupby(['group'])

# determine the number of levels for each pest
df4 = pest_tax_group.size()
df5 = df4.to_frame()
df5.rename(columns={0:'group_size'}, inplace=True)

# create a column with the number of the group
df5.reset_index(inplace=True)

# merge the group with the taxonomy
df_tax1 = pd.merge(df_tax, df5, left_on=['group'], right_on=['group'])

# make a pivot for taxonomy, when levels are equal to 7
df_tax2 = df_tax1[df_tax1['group_size'] == 7]
df_tax3 = df_tax2.pivot(index='group', columns='level', values=['eppocode', 'prefname'])

# rename columns
df_tax3.rename(columns={1:'kingdom', 2:'phylum', 3:'class', 4:'order', 5:'family', 6:'genus', 7:'species'}, 
               inplace=True)
df_tax3.columns = [' '.join(col).strip() for col in df_tax3.columns.values]

# add columns for missing taxon ranks
df_tax3.insert(loc = 2, column = 'eppocode subphylum', value = '')
df_tax3.insert(loc = 10, column = 'prefname subphylum', value = '')
df_tax3.insert(loc = 5, column = 'eppocode suborder', value = '')
df_tax3.insert(loc = 14, column = 'prefname suborder', value = '')

# make a pivot for taxonomy, when levels are equal to 8
df_tax2 = df_tax1[df_tax1['group_size'] == 8]
df_tax4 = df_tax2.pivot(index='group', columns='level', values=['eppocode', 'prefname'])
df_tax4.rename(columns={1:'kingdom', 2:'phylum', 3:'subphylum', 4:'class', 5:'order', 6:'family', 
                        7:'genus', 8:'species'}, 
               inplace=True)
df_tax4.columns = [' '.join(col).strip() for col in df_tax4.columns.values]
df_tax4.insert(loc = 5, column = 'eppocode suborder', value = '')
df_tax4.insert(loc = 14, column = 'prefname suborder', value = '')

# make a pivot for taxonomy, when levels are equal to 9
df_tax2 = df_tax1[df_tax1['group_size'] == 9]
df_tax5 = df_tax2.pivot(index='group', columns='level', values=['eppocode', 'prefname'])
df_tax5.rename(columns={1:'kingdom', 2:'phylum', 3:'subphylum', 4:'class', 5:'order', 6:'suborder', 
                        7:'family', 8:'genus', 9:'species'}, 
               inplace=True)
df_tax5.columns = [' '.join(col).strip() for col in df_tax5.columns.values]
df_tax5.columns

frames = [df_tax3, df_tax4, df_tax5]
tax_table = pd.concat(frames)

# save to csv
tax_table.to_csv('../process_data/tax_table.csv')

tax_table


# merge taxonomy with crop pests and remove duplicates
result = pd.merge(tax_table, pest_codes, left_on=['eppocode species'], right_on=['eppocode'], how='left')
result = result.drop_duplicates()

# save to csv
result.to_csv('../process_data/eppo_crop_pest_data.csv')

## 4. Create a list of pests to present to the farmer

Get pest common names from EPPO to be present to the farmer, so that he can confirm that the pest in present in the region.

pests = result['eppocode species'].unique()

%%time
# Get common name for each pest. This can take 4 minutes to run for a list of 2000 pest codes

service = '/names'
df = pd.DataFrame()


concatenated_dfs = []

for pest in pests:
    url = base_url+pest+service+'?authtoken='+token
    # print(url)
    response = requests.get(url)
    data = response.json()
    if data != None:
        species = data[0]['fullname']
        print(f'Species:. {species}')
        if len(data) > 1:
            common = data[1]['fullname']
        else:
            common = ""
        print(f'Common: {common}')
        values = {'pest_eppocode': [pest], 'scientific_name': [species], 
                   'common_name': [common]}
        df = pd.DataFrame.from_dict(values)
        concatenated_dfs.append(df)

df_common = pd.concat(concatenated_dfs)
df_common.head()

# save to csv
df_common.to_csv('../process_data/pest_common.csv')
