==> notebooks/05_determine_predation.ipynb <==
# 05 - Create list of pests, crop and bird of the selected area

### Description

The goal of this notebook is to:
- determine if birds prey the pests identified for the crops in the area

### Inputs

- list of bird species from the occurrence table, prepared in step 03
- list of pests for the crops in the selected region, prepared in step 04

### External resources 

- API service of Global Biotic Interactions https://github.com/globalbioticinteractions/globalbioticinteractions/wiki/API#interactions

### Outputs

- List of pests' taxonomy, the corresponding hosting crop, and bird species, with indication if they eat that pest

# import modules
import pandas as pd
import requests
%matplotlib inline

## 1. Read list of bird species

We will open the list of species occurrences created in step 03 of the workflow. The table also contains columns with the trait data. We will filter to obtain only bird species that use crops as feeding areas.

# load eBird selected occurrence data
occ_file = '../process_data/eBird_sel.csv'

pdf_o = pd.read_csv(occ_file, low_memory=False)

bird_data_name = pdf_o

# filter species with traits classified, to obtain species that use crops as feeding areas
options = [0, 1]
bird_data_name = bird_data_name[bird_data_name['Annual_crops'].isin(options)]
bird_data_name.head()

# list of species occurrences
#in_file = '../process_data/eBird_sel.csv'
#df_birds = pd.read_csv(in_file, low_memory=False)
df_unique_birds = bird_data_name[['TAXON CONCEPT ID_x', 'SCIENTIFIC NAME_x', 'ORDER1']].drop_duplicates()
df_unique_birds

## 2. Read list of pests and combine with bird species

We will open the list of pests created at step 04. The pest is considered at order taxnonomic level. Afterwards, combine with species to obtain all possible combinations.

# read list of pests for the crops existing in the area
# set the taxon rank of the pest at the list of columns of df_pests

in_file1 = '../process_data/eppo_crop_pest_data.csv'
eppo_crop_pest_data = pd.read_csv(in_file1, low_memory=False)
df_pests = eppo_crop_pest_data[['eppocode species', 'prefname order', 'crop_eppocode']].drop_duplicates()
df_pests_order = df_pests

df_pests_order

# Combine data between the two dataframes
cross_data = pd.merge(df_unique_birds, df_pests_order, how='cross')
bird_prey = cross_data[['TAXON CONCEPT ID_x', 'SCIENTIFIC NAME_x', 'prefname order', 'eppocode species']].drop_duplicates()

bird_prey

There is a total of 87648 possible combinations between bird species, pests and crops.

Create a list with unique combinations of birds and possible pest preys:

bird_prey_unique = bird_prey[['SCIENTIFIC NAME_x', 'prefname order']].drop_duplicates()
bird_prey_unique

print('Number of bird species:',len(bird_prey_unique['SCIENTIFIC NAME_x'].unique()))
print('Number of pest orders:',len(bird_prey_unique['prefname order'].unique()))

There are 1364 combinations between 124 bird species and 11 pest orders.

## 3. Get interaction data from GLOBI
 
- Request data from GLOBI API to determine if birds prey on pests.

'''
# define a function to get data from GLOBI.
# Returns 0 if it is not eaten, 1 otherwise
# inputs
   - predator: bird species
   - prey: pest order
'''
        
base_url = 'https://api.globalbioticinteractions.org/taxon/'
interaction = '/eats/'

def get_globi(pred, prey):
    url = base_url+pred+interaction+prey
    print(url)
    response = requests.get(url)
    data = response.json()
    if len(data['data']) > 0:
        return 1
    else:
        return 0 

%time
# Get data from GLOBI about birds eating preys


# add bolean column to indicate if bird eats pest
bird_prey_unique['eats'] = bird_prey_unique.apply(lambda x: get_globi(x['SCIENTIFIC NAME_x'], \
                                                                      x['prefname order']), axis=1)
bird_prey_unique.head()

# save to file
bird_prey_unique.to_csv('../process_data/bird_prey_unique.csv')

# merge the new classification with the table of the taxonomic classification of the pests
bird_prey_crop = pd.merge(eppo_crop_pest_data, bird_prey_unique, left_on=['prefname order'], right_on=['prefname order'],  how='left')

# save to file
bird_prey_crop.to_csv('../process_data/bird_prey_crop.csv')
bird_prey_crop

The table above contains the taxonomic classification of the pests, in EPPO code format (eppocode_?) and taxon name (prefname_?), the EPPO code of the crop (crop_eppocode), the bird species and the boolean column `eats`, which indicates if the bird species eats or not the pest.
