==> notebooks/07_calculate_DFI.ipynb <==
# 07 - Calculate daily food intake by birds

#### Description

The goal of this notebook is to calculate the Daily Food Intake (DFI) by birds.

It implements the following steps:

- determine the Daily Energy Expenditure (DEE) for each bird species
- determine the Daily Food Intake (DFI)

#### Inputs

- table with crop fraction for which birds provide service, created in step 06: `eBird_area_data_percent-step3.csv`
- table moisture and energy content of invertebrates, compiled from Crocker et al. (2002)
- table with assimilation efficiency of birds, compiled from Crocker et al. (2002)
   

#### Outputs

- table with DFI and consumption at the nesting/growing season

## 1. Calculate bird Daily Food Intake (DFI)

To calculate DFI, we will use the following expression (Crocker et al, 2002):
```
Daily foor intake (g) = DEE (kj) / Energy in Food (kj/g) x (1 - Moisture) x Assimilation Efficiency
```
where DEE is the Daily Energy Expenditure, which is calculated by the following:
```
Log(DEE) = Log a + b × (log Body weight).
loga = 1.0220
b = 0.6745
```


### 1.1 Prepare tables

The determination of DFI is done in steps:

- load data from csv files, to save time (avoid repeating steps 1 to 4)
- calculate DEE
- calculate assimilation efficiency
- prepare tables, to gain efficiency
    - explode crop data, to create simple tables for temporary and for permanent crops.

# import modules
import pandas as pd
import numpy as np
from pandas import DataFrame
import geopandas as gpd
from pyproj import Proj, CRS,transform
import matplotlib.pyplot as plt

Read data with the fraction that each crop represents in the buffer area:

# read table of occurrences containing crop fraction in the buffer

bird_data_name = pd.read_csv('../process_data/eBird_area_data_percent-step3.csv', low_memory = False)

# increase the numebr of columns that display
pd.set_option('display.max_columns', None)

# function to calculate DEE, uses values for All Birds in DEFRA

def calc_dee(mass):
    intercept = 1.0220
    coef = 0.6745
    logDEE = intercept + (coef * np.log10(mass))
    dee = np.power(logDEE, 10)
    return dee
    

Add Daily Energy Expenditure (DEE) for each species to the occurrence table:

# add DEE to the bird occurrence table
bird_data_name['DEE'] = bird_data_name.apply(lambda x: calc_dee(x['Mass']), axis=1)

We will read a table containing the moisture percentage, energy and fresh weight of invertebrates. This table was extracted from Crocker et al. (2002).

# load energy and moisture table 
energy_moist = pd.read_csv('../external_data/DEFRA/energy_moisture.csv')
energy_moist.head()

We will read a table containing the assimilation efficiency of several bird orders. This table was extracted from Crocker et al. (2002).

# load assimilation table
assim_eff = pd.read_csv('../external_data/DEFRA/assimilation_efficienc.csv')
assim_eff.head()

Combine both tables, in order to combinate of energy and moisture content of invertebrates with assimilation efficiency of birds.

# merge energy table with assimilation table
energy_food = pd.merge(energy_moist, assim_eff, how='cross')

# calculate energy in food
energy_food['energy_food'] = energy_food['fresh_weight_kj_g'] * energy_food['assim_eff']
energy_food

# rename columns, to facilitate merge
energy_food.rename(columns={"eppocode_order": "eppo_pest", "order": "Order2"}, inplace=True)

%%time
# ensure that columns of crop_pest are lists. Can take 1 min to run.

import ast

bird_data_name['consum_t_ord'] = bird_data_name['consum_t_ord'].apply(ast.literal_eval)
bird_data_name['consum_p_ord'] = bird_data_name['consum_p_ord'].apply(ast.literal_eval) 

bird_data_name.columns



Next steps are to prepare table structure so that calculations of DFI can be done. This needs to be done separately for temporary and permanent crops. 

# explode the dataframe to create new rows, create table for temporary crops

table_df_t = bird_data_name[['GLOBAL UNIQUE IDENTIFIER', 'Order2', 'Avibase.ID2', 'SAMPLING EVENT IDENTIFIER', \
                             'SCIENTIFIC NAME_x', 'consum_t_ord', 'DEE', 'Annual_crops', 'Permanent_crops', \
                             'Proportion_invertebrates_diet']]
table_df_t = table_df_t.explode('consum_t_ord')


table_df_t

# funtion to split crops and pests
def split_values(pair):
    if type(pair) != list: # wrong case, returns nan
        a = np.nan
        b = np.nan
    else:
        a = pair[0]
        b = pair[1]
    return a, b

%%time
# split crops and pests in two columns, temporary crops. Can take 40 sec to run
table_df_t['eppo_crop'], table_df_t['eppo_pest'] = \
zip(*table_df_t.apply( lambda x: split_values(x['consum_t_ord']), axis=1))

# remove old column, to save memory space
table_df_t.drop(columns=['consum_t_ord'], inplace=True)
table_df_t

# explode the dataframe to create new rows, create table for permanent crops

table_df_p = bird_data_name[['GLOBAL UNIQUE IDENTIFIER', 'Order2', 'Avibase.ID2', 'SAMPLING EVENT IDENTIFIER', \
                             'SCIENTIFIC NAME_x', 'consum_p_ord', 'DEE', 'Annual_crops', 'Permanent_crops', \
                             'Proportion_invertebrates_diet']]
table_df_p = table_df_p.explode('consum_p_ord')


%%time
# split crops and pests in two columns, permanent crops
table_df_p['eppo_crop'], table_df_p['eppo_pest'] = \
zip(*table_df_p.apply( lambda x: split_values(x['consum_p_ord']), axis=1))

# remove old column, to save memory space
table_df_p.drop(columns=['consum_p_ord'], inplace=True)
table_df_p

### 1.2. Calculate DFI

The DFI is the ratio between the DEE and energy in food, in this case, pests. We will calculate that for temporary and permanent crops. In the end, both values will combined, weighted by the proportion that each type of crop represents in the point buffer.

# merge tables with energy_food

table_df_t = pd.merge(table_df_t, energy_food, on=['Order2','eppo_pest'])
table_df_p = pd.merge(table_df_p, energy_food, on=['Order2','eppo_pest'])

# calculate DFI
table_df_t['DFI'] = table_df_t['DEE'] / table_df_t['energy_food']
table_df_p['DFI'] = table_df_p['DEE'] / table_df_p['energy_food']

table_df_p

# add column with crop type

table_df_t['crop_type'] = 'temp'
table_df_p['crop_type'] = 'perm'

table_df_p.columns

#concatenate the two tables
table_DFI = pd.concat([table_df_t, table_df_p])

# drop columns not needed anymore
table_DFI.drop(columns=['ID', 'Energy_food_kj_g', 'moisture_perc', 'fresh_weight_kj_g', 'assim_eff', 'energy_food'], inplace=True)


table_DFI = table_DFI.drop_duplicates()

table_DFI

table_DFI.to_csv('../process_data/table_DFI_1.csv')

### 1.3. Determine how many pests each species eats, to determine fractions

Calculate number of references, i.e., how many pests can each species eat, per species list. Serves to determine the fraction of 
the diet that applies to each pest.

# read table that resulted from 1.1 and 1.2
table_DFI = pd.read_csv('../process_data/table_DFI_1.csv', low_memory=False)

# select columns to use for the analysis

df = table_DFI[['SCIENTIFIC NAME_x', 'Avibase.ID2', 'SAMPLING EVENT IDENTIFIER', 'eppo_pest']]

df2 = df.drop_duplicates()

# create a table with the count number of pests consumed by each bird species in each list
df2['count_eat_ref'] = df2.groupby(['SCIENTIFIC NAME_x', 'SAMPLING EVENT IDENTIFIER'])['eppo_pest'].transform('count')

# remove duplicates 
df3 = df2[['Avibase.ID2', 'SAMPLING EVENT IDENTIFIER', 'count_eat_ref']].drop_duplicates()


# merge with DFI table
table_DFI = pd.merge(table_DFI, df3, left_on=['Avibase.ID2', 'SAMPLING EVENT IDENTIFIER'], right_on=['Avibase.ID2', 'SAMPLING EVENT IDENTIFIER'])

# calculate fraction of consumption per pest, by dividing by the count of preyed pests
# Also considers the proportion of the diet
table_DFI['frac_DFI'] = table_DFI['DFI'] / table_DFI['count_eat_ref'] * table_DFI['Proportion_invertebrates_diet']

# calculate total consumption per season, by multiplying by the number of days of the season (01st Apr and 30th Jun)
table_DFI['consum_season'] = table_DFI['frac_DFI'] * 91

table_DFI

table_DFI.to_csv('../process_data/table_DFI_2.csv')
